requirements.txt
black
datetime
pylint
pysimplegui
pyyaml
telethon
yamllint
